categorize_links/requirements.txt
google-auth-oauthlib
google-api-python-client
langchain
openai
python-dotenv